+++ CHANGELOG.rst
@@ -15,13 +15,14 @@ Subsections for each version can be one of the following;
 - ``Security`` in case of vulnerabilities.
 
 Each individual change should have a link to the pull request after the description of the change.
-
 1.3.1 (unreleased)
 ------------------
 
 Changed
 ^^^^^^^
 - Refactored NominalToIntegerTransformer tests in new format `#261 <https://github.com/lvgig/tubular/pull/261>`_
+- DatetimeInfoExtractor.mappings_provided changed from a dict.keys() object to list so transformer is serialisable. `#258 <https://github.com/lvgig/tubular/pull/258>`_
+- Created BaseNumericTransformer class to support test refactor of numeric file
 
 1.3.0 (2024-06-13)
 ------------------